// bindings/typescript/api/src/raw/enums.ts
export const MLN_ANIMATION_OPTION_FIELD = {
  MLN_ANIMATION_OPTION_DURATION: 1,
  MLN_ANIMATION_OPTION_VELOCITY: 2,
  MLN_ANIMATION_OPTION_MIN_ZOOM: 4,
  MLN_ANIMATION_OPTION_EASING: 8,
  MLN_ANIMATION_OPTION_TRANSITION_ID: 16,
} as const;
// ...
export const MLN_CAMERA_OPTION_FIELD = {
  MLN_CAMERA_OPTION_CENTER: 1,
  MLN_CAMERA_OPTION_ZOOM: 2,
  MLN_CAMERA_OPTION_BEARING: 4,
  MLN_CAMERA_OPTION_PITCH: 8,
  MLN_CAMERA_OPTION_CENTER_ALTITUDE: 16,
  MLN_CAMERA_OPTION_PADDING: 32,
  MLN_CAMERA_OPTION_ANCHOR: 64,
  MLN_CAMERA_OPTION_ROLL: 128,
  MLN_CAMERA_OPTION_FOV: 256,
} as const;